refactor: 공통 레이아웃 변경 및 기존 page.tsx 삭제 (#28)
