+++ src/app/(home)/layout.tsx
@@ -1,0 +1,48 @@
+import '@/src/styles/tailwind.css';  // Tailwind CSS가 포함된 파일
+import '@/src/styles/globals.css';   // 글로벌 스타일을 포함한 파일
+import Header from '@/src/components/layout/Header'; // 공통 헤더 컴포넌트 임포트
+
+export const metadata = {
+    title: 'Next.js 프로젝트',
+    description: 'Generated by Next.js',
+};
+
+export default function HomeLayout({
+                                       children,
+                                   }: {
+    children: React.ReactNode;
+}) {
+    return (
+        <html lang="ko">
+        <head>
+            <meta charSet="UTF-8" />
+            <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+            <meta httpEquiv="X-UA-Compatible" content="ie=edge" />
+            <title>{metadata.title}</title>
+            <meta name="description" content={metadata.description} />
+        </head>
+        <body>
+        {/* 공통 헤더 내용 */}
+        <Header />
+
+        {/* 메인 콘텐츠 */}
+        <main className="max-w-[1200px] min-h-[900px] mx-auto p-4"> {/* 최대 너비 1200px 고정 */}
+            {children}
+        </main>
+
+        {/* 공통 푸터 내용 */}
+        <footer className="bg-white text-gray-5 py-6 text-center border-t border-gray-200">
+            <div className="flex flex-col items-center gap-2">
+                <div className="flex gap-6 text-sm">
+                    <a href="#" className="hover:underline">이용약관</a>
+                    <a href="#" className="hover:underline">개인정보처리방침</a>
+                    <a href="#" className="hover:underline">라이선스 준수</a>
+
+                    <p className="text-sm">고객문의 : <a href="mailto:your-email@example.com" className="text-blue-500 hover:underline">your-email@example.com</a></p>
+                </div>
+            </div>
+        </footer>
+        </body>
+        </html>
+    );
+}
